requirements.txt
streamlit
pandas
pdfplumber
python-dotenv
dateparser
